Refactor: Extract code related to document processing and formatting into separate files implementing IDocumentProcessor
