+++ XamlStyler.Service/DocumentProcessors/IDocumentProcessor.cs
@@ -1,0 +1,10 @@
+﻿using System.Text;
+using System.Xml;
+
+namespace XamlStyler.Core.DocumentProcessors
+{
+    internal interface IDocumentProcessor
+    {
+        void Process(XmlReader xmlReader, StringBuilder output, ElementProcessContext elementProcessContext);
+    }
+}
